delete globalscheduler_test, update scheduler_test
